requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
streamlit
requests
langchain